Added univariate selection to base class
